moved power-triangle function to visu, added some more examples
